implement basic plugin customization
